Cleaning the dependencies and adding more aliases

+++ mac/clean.sh
@@ -1,15 +1,17 @@
 #!/bin/sh
 
+$(brew --prefix)/opt/fzf/uninstall >>/dev/null
+
 # Personal scripts and configuration
 rm -rf ${HOME}/.config/mbb
 
 # Common configuration files
-rm ${HOME}/.zshrc
-rm ${HOME}/.nanorc
-rm ${HOME}/.p10k.zsh
-rm ${HOME}/.fzf.zsh
-rm ${HOME}/.fzf.bash
-rm ${HOME}/.zshrc.pre-oh-my-zsh
+[ -f ${HOME}/.zshrc ] && rm ${HOME}/.zshrc
+[ -f ${HOME}/.nanorc ] && rm ${HOME}/.nanorc
+[ -f ${HOME}/.p10k.zsh ] && rm ${HOME}/.p10k.zsh
+[ -f ${HOME}/.fzf.zsh ] && rm ${HOME}/.fzf.zsh
+[ -f ${HOME}/.fzf.bash ] && rm ${HOME}/.fzf.bash
+[ -f ${HOME}/.zshrc.pre-oh-my-zsh ] && rm ${HOME}/.zshrc.pre-oh-my-zsh
 
 # Common folders for terminal plugins
 rm -rf ${HOME}/.oh-my-zsh
@@ -27,5 +29,5 @@ rm -rf "${HOME}/.git-fuzzy"
 # rm ${HOME}/.gitconfig
 
 # Third party tools
-rm ${HOME}/.ansiweatherrc
-rm ${HOME}/.ticker.yaml
+[ -f ${HOME}/.ansiweatherrc ] && rm ${HOME}/.ansiweatherrc
+[ -f ${HOME}/.ticker.yaml ] && rm ${HOME}/.ticker.yaml
